Działająca klasa HouseholdAgent
