+++ MicrogridEMS/MicrogridEMS.py
@@ -1,10 +1,15 @@
 import numpy as np
 from numpy import *
 import dataoperation as do
-import Optimizer as ems
+from HouseholdAgent import HouseholdAgent
 
-optimizer = ems.Optimizer()
-optimizer.calculate()
+
+agent = HouseholdAgent()
+agent.optimize()
+
+do.printcolumn(agent.offers.res_m.power)
+
+
